requirements.txt
pytest==7.4.3
playwright==1.40.0
allure_python_commons==2.13.5

page~=0.2
